requirements.txt
scipy
numpy
statsmodels
boto3
BPTK-Py
sklearn
pyzmq>=17
tornado
jupyterlab
elasticsearch
keras
tensorflow
nodejs